+++ packages/expo-updates/android/src/main/java/expo/modules/updates/loader/Loader.kt
@@ -1,7 +1,6 @@
 package expo.modules.updates.loader
 
 import android.content.Context
-import android.util.Log
 import expo.modules.updates.UpdatesConfiguration
 import expo.modules.updates.UpdatesUtils
 import expo.modules.updates.db.UpdatesDatabase
@@ -10,6 +9,8 @@ import expo.modules.updates.db.entity.UpdateEntity
 import expo.modules.updates.db.enums.UpdateStatus
 import expo.modules.updates.loader.FileDownloader.AssetDownloadCallback
 import expo.modules.updates.loader.FileDownloader.RemoteUpdateDownloadCallback
+import expo.modules.updates.logging.UpdatesErrorCode
+import expo.modules.updates.logging.UpdatesLogger
 import expo.modules.updates.manifest.ManifestMetadata
 import expo.modules.updates.manifest.Update
 import java.io.File
@@ -26,6 +27,7 @@ import java.util.*
 abstract class Loader protected constructor(
   protected val context: Context,
   private val configuration: UpdatesConfiguration,
+  protected val logger: UpdatesLogger,
   private val database: UpdatesDatabase,
   private val updatesDirectory: File,
   private val loaderFiles: LoaderFiles
@@ -132,10 +134,8 @@ abstract class Loader protected constructor(
 
   private fun finishWithSuccess() {
     if (callback == null) {
-      Log.e(
-        TAG,
-        this.javaClass.simpleName + " tried to finish but it already finished or was never initialized."
-      )
+      val cause = Exception("Null callback in finishWithSuccess")
+      logger.error("${this.javaClass.simpleName} tried to finish but it already finished or was never initialized.", cause, UpdatesErrorCode.UpdateFailedToLoad)
       return
     }
 
@@ -156,12 +156,9 @@ abstract class Loader protected constructor(
   }
 
   private fun finishWithException(e: Exception) {
-    Log.e(TAG, e.message!!)
+    logger.error("Load error", e, UpdatesErrorCode.UpdateFailedToLoad)
     if (callback == null) {
-      Log.e(
-        TAG,
-        this.javaClass.simpleName + " tried to finish but it already finished or was never initialized."
-      )
+      logger.error("${this.javaClass.simpleName} tried to finish but it already finished or was never initialized.", e, UpdatesErrorCode.UpdateFailedToLoad)
       return
     }
     callback!!.onFailure(e)
@@ -189,11 +186,8 @@ abstract class Loader protected constructor(
     // but different scope keys, we should try to launch something rather than show a cryptic
     // error to the user.
     if (existingUpdateEntity != null && existingUpdateEntity.scopeKey != newUpdateEntity.scopeKey) {
+      logger.warn("Loaded an update with the same ID but a different scopeKey than one we already have on disk. This is a server error. Overwriting the scopeKey and loading the existing update.")
       database.updateDao().setUpdateScopeKey(existingUpdateEntity, newUpdateEntity.scopeKey)
-      Log.e(
-        TAG,
-        "Loaded an update with the same ID but a different scopeKey than one we already have on disk. This is a server error. Overwriting the scopeKey and loading the existing update."
-      )
     }
 
     if (existingUpdateEntity != null && existingUpdateEntity.status == UpdateStatus.READY) {
@@ -258,7 +252,7 @@ abstract class Loader protected constructor(
             } else {
               "key " + assetEntity.key
             }
-            Log.e(TAG, "Failed to download asset with $identifier", e)
+            logger.error("Failed to download asset with $identifier", e)
             handleAssetDownloadCompleted(assetEntity, AssetLoadResult.ERRORED)
           }